requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
email_validator==2.1.0.post1

# Task Queue
celery==5.3.6
redis==5.0.1

# LLM Integration
openai==1.3.0
anthropic==0.5.0

# Payment Processing
stripe==7.8.0

# Security & Rate Limiting
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Flask-CORS==4.0.0

# Caching
Flask-Caching==2.1.0

# Utilities
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0

# Database
psycopg2-binary==2.9.9

# Production Server
gunicorn==21.2.0

# Excel Export
openpyxl==3.1.2
pandas==2.1.4

# Email
Flask-Mail==0.9.1

# Scheduling
APScheduler==3.10.4
